Iss33 (#34)
* Removing aritificial river segments

This PR assures that no artificial river segments are added to the
downstream of outlet river segments. This was previous enable to
assure WatRoute routing throughout the entire river network, as
WatRoute does not route runoff for the outlet segments.

Reported-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>
Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>

* Removing aritificial river segments

This PR assures that no artificial river segments are added to the
downstream of outlet river segments. This was previous enable to
assure WatRoute routing throughout the entire river network, as
WatRoute does not route runoff for the outlet segments.

Reported-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>
Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>

---------

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>