Defining and testing cutsom parameter inclusion
This PR introduces a simple mechanism to include parameters only
if the relevant processes are included in the model configuration.

The `src/meshflow/templating/MESH_parameters_hydrology.ini.jinja`
file defines the parmaters based on select number of options included
in the configuration of a MESH instance.

The main `MESHWorkflow` class is consequently adjusted to be able
to deal with such definitions. If users enter specific options
using the `settings.run_options`, such adjustments would be implemented.
Otherwise, parameters related to default configuration will be
templated.

The test example is updated to reflect such capability.

Reported-by: Daniel Princz <daniel.princz@ec.gc.ca>
Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>